+++ src/healthcheck.service.js
@@ -1,0 +1,6 @@
+const send = (req, res) => {
+    res.send('<h2>The Slash Command and Dialog app is running</h2> <p>Follow the' +
+        ' instructions in the README to configure the Slack App and your environment variables.</p>');
+};
+
+module.exports = { send };
